Create utils file and  remove duplication
