+++ src/UpwardIntervalExercise.js
@@ -3,19 +3,10 @@ import PropTypes from 'prop-types';
 import Tone from 'tone';
 import Teoria from 'teoria';
 
-import AnswerButtons, { POSSIBLE_INTERVALS } from './AnswerButtons.js';
+import AnswerButtons from './AnswerButtons.js';
 import PlayButton from './PlayButton.js';
 import InstructionsText from './InstructionsText.js'
 
-
-function getRandomInt(max) {
-    return Math.floor(Math.random() * Math.floor(max));
-}
-
-
-export function generateRandomUpwardInterval() {
-    return POSSIBLE_INTERVALS[getRandomInt(POSSIBLE_INTERVALS.length)]
-}
 
 class UpwardIntervalExercise extends Component {
     constructor() {
